refactor leastsq into complexity
